formatting and type changes on Card sizes

+++ Bingo.Library/Card.cs
@@ -4,7 +4,7 @@ public class Card : ICard
 {
     public byte Columns { get; init; }
     public byte Rows { get; init; }
-    public short TotalSquares { get; init; }
+    public byte TotalSquares { get; init; }
     public byte BaseSquareValue { get; init; }
     public int RowValueOffset { get; init; }
     public byte BonusColumns { get; init; }
@@ -19,7 +19,7 @@ public class Card : ICard
         RowValueOffset = rowOffsetValue;
         BonusColumns = bonusColumns;
         BonusMultiplier = bonusMultiplier;
-        TotalSquares = (short)(columns * rows);
+        TotalSquares = (byte)(columns * rows);
         BonusSkipChar = 'P';
     }
 }
