+++ Bingo.Library/ICard.cs
@@ -4,7 +4,7 @@ public interface ICard : IGrid
 {
     public byte Columns { get; init; }
     public byte Rows { get; init; }
-    public short TotalSquares { get; init; }
+    public byte TotalSquares { get; init; }
     public byte BaseSquareValue { get; init; }
     public int RowValueOffset { get; init; }
     public byte BonusColumns { get; init; }
